refactor(string_support.rb): `String#camerize` / `String#underscore`
