+++ lib/fusuma/string_support.rb
@@ -1,0 +1,16 @@
+# support camerize and underscore
+class String
+  def camelize
+    split('_').map do |w|
+      w.capitalize
+    end.join
+  end
+
+  def underscore
+    gsub(/([A-Z]+)([A-Z][a-z])/, '\1_\2')
+      .gsub(/([a-z\d])([A-Z])/, '\1_\2')
+      .gsub('::', '/')
+      .tr('-', '_')
+      .downcase
+  end
+end
